+++ std-udp-recv/src/common/FrameStats.cpp
@@ -7,7 +7,7 @@ using namespace std;
 using namespace chrono;
 
 FrameStats::FrameStats(string detector_name, const int module_id,
-                       const int n_packets_per_frame, const size_t stats_time) :
+                       const size_t n_packets_per_frame, const size_t stats_time) :
             detector_name_(move(detector_name)),
             module_id_(module_id),
             n_packets_per_frame_(n_packets_per_frame),
